requirements.txt
PyAutoGUI==0.9.54
Pillow==10.0.1
pytesseract==0.3.10
opencv-python==4.8.1.78
requests==2.31.0